proto: use decaf377 signature proto types instead of bytes
