+++ crates/proto/src/gen/penumbra.core.component.shielded_pool.v1alpha1.serde.rs
@@ -1998,7 +1998,7 @@ impl serde::Serialize for SpendBody {
         if self.nullifier.is_some() {
             len += 1;
         }
-        if !self.rk.is_empty() {
+        if self.rk.is_some() {
             len += 1;
         }
         let mut struct_ser = serializer.serialize_struct("penumbra.core.component.shielded_pool.v1alpha1.SpendBody", len)?;
@@ -2008,9 +2008,8 @@ impl serde::Serialize for SpendBody {
         if let Some(v) = self.nullifier.as_ref() {
             struct_ser.serialize_field("nullifier", v)?;
         }
-        if !self.rk.is_empty() {
-            #[allow(clippy::needless_borrow)]
-            struct_ser.serialize_field("rk", pbjson::private::base64::encode(&self.rk).as_str())?;
+        if let Some(v) = self.rk.as_ref() {
+            struct_ser.serialize_field("rk", v)?;
         }
         struct_ser.end()
     }
@@ -2097,16 +2096,14 @@ impl<'de> serde::Deserialize<'de> for SpendBody {
                             if rk__.is_some() {
                                 return Err(serde::de::Error::duplicate_field("rk"));
                             }
-                            rk__ = 
-                                Some(map_.next_value::<::pbjson::private::BytesDeserialize<_>>()?.0)
-                            ;
+                            rk__ = map_.next_value()?;
                         }
                     }
                 }
                 Ok(SpendBody {
                     balance_commitment: balance_commitment__,
                     nullifier: nullifier__,
-                    rk: rk__.unwrap_or_default(),
+                    rk: rk__,
                 })
             }
         }
