+++ proto/penumbra/penumbra/core/component/governance/v1alpha1/governance.proto
@@ -90,7 +90,7 @@ message DelegatorVoteBody {
   // The nullifier of the input note.
   sct.v1alpha1.Nullifier nullifier = 6;
   // The randomized validating key for the spend authorization signature.
-  bytes rk = 7;
+  crypto.decaf377_rdsa.v1alpha1.SpendVerificationKey rk = 7;
 }
 
 message DelegatorVoteView {
