+++ proto/penumbra/penumbra/core/component/shielded_pool/v1alpha1/shielded_pool.proto
@@ -82,7 +82,7 @@ message SpendBody {
   // The nullifier of the input note.
   sct.v1alpha1.Nullifier nullifier = 6;
   // The randomized validating key for the spend authorization signature.
-  bytes rk = 4;
+  crypto.decaf377_rdsa.v1alpha1.SpendVerificationKey rk = 4;
 }
 message SpendView {
   message Visible {
