+++ proto/penumbra/penumbra/core/keys/v1alpha1/keys.proto
@@ -41,10 +41,6 @@ message SpendKey {
   bytes inner = 1;
 }
 
-message SpendVerificationKey {
-  bytes inner = 1;
-}
-
 message FullViewingKey {
   bytes inner = 1;
 }
