+++ proto/penumbra/penumbra/crypto/decaf377_rdsa/v1alpha1/decaf377_rdsa.proto
@@ -8,3 +8,7 @@ message SpendAuthSignature {
 message BindingSignature {
   bytes inner = 1;
 }
+
+message SpendVerificationKey {
+  bytes inner = 1;
+}
